refactor(issue-33): just go with a single top level internal package
