+++ cmd/griot/BUILD.bazel
@@ -8,7 +8,7 @@ go_library(
     visibility = ["//visibility:private"],
     deps = [
         "//cmd/griot/app",
-        "//cmd/internal/command",
+        "//internal/command",
     ],
 )
