+++ cmd/griot/content/BUILD.bazel
@@ -7,6 +7,6 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//cmd/griot/content/upload",
-        "//cmd/internal/command",
+        "//internal/command",
     ],
 )
